React-TodoList
This project is a simple todo list application built with React. It demonstrates the usage of React hooks, state management, and event handling.

Technologies used:
- React
- CSS

To use the application, simply type your todo item in the input field and press the "Add" button. Click on a todo item to remove it from the list.